Add website image (photo03.png)
Uploaded photo03.png for website usage.
This image is intended for display on the homepage or a specific section of the site.
Explanation:

This provides more context for the change. It clarifies the purpose of the uploaded image and where it's intended to be used on the website. You can modify the second sentence if you have a more specific location or purpose for photo3 .png.